+++ src/onegov.org/onegov/org/models/organisation.py
@@ -49,6 +49,7 @@ class Organisation(Base, TimestampMixin):
     homepage_cover = meta_property('homepage_cover')
     bank_account = meta_property('bank_account')
     bank_beneficiary = meta_property('bank_beneficiary')
+    bank_payment_order_type = meta_property('bank_payment_order_type')
 
     @contact.setter
     def contact(self, value):